Fix an ambiguous exception call that only shows up off MSVC
OutOfMemoryException takes both an i64 and a u64 size, and the initializer
list constructor passed it std::initializer_list::size(), which is size_t.
On Linux size_t is unsigned long while u64 is unsigned long long and i64 is
long long, so converting to either overload is an equal rank integral
conversion and neither wins. On MSVC size_t is unsigned long long, matching
u64 exactly, so it compiled there and nowhere else.

Cast the size explicitly. The other call sites pass Opal's own u64 size
types, whose greater rank already makes the product a u64, so they were
never ambiguous.

+++ include/opal/container/hash-map.h
@@ -381,13 +381,13 @@ Opal::HashMap<KeyType, ValueType>::HashMap(std::initializer_list<pair_type> pair
 {
     if (Reserve(pairs.size()) != ErrorCode::Success)
     {
-        throw OutOfMemoryException(m_allocator->GetName(), pairs.size() * sizeof(pair_type));
+        throw OutOfMemoryException(m_allocator->GetName(), static_cast<u64>(pairs.size() * sizeof(pair_type)));
     }
     for (const auto& pair : pairs)
     {
         if (Insert(Opal::Clone(pair.key, m_allocator), Opal::Clone(pair.value, m_allocator)) != ErrorCode::Success)
         {
-            throw OutOfMemoryException(m_allocator->GetName(), pairs.size() * sizeof(pair_type));
+            throw OutOfMemoryException(m_allocator->GetName(), static_cast<u64>(pairs.size() * sizeof(pair_type)));
         }
     }
 }
